chapter10/creating_a_complex_architecture_on_azure_with_ai_services/requirements.txt
flask
openai
azure-identity
